% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_covid_plot.R
\name{bb_covid_plot}
\alias{bb_covid_plot}
\title{Plot government budget balance during the COVID-19 pandemic}
\usage{
bb_covid_plot()
}
\value{
plot object
}
\description{
This function plots the government budget balance during the years of the COVID-19 pandemic.
Since the pandemic started at the beginning of 2022, the budget balance for the years 2020, 2021, and 2022 is plotted.
}
\examples{

bb_covid_plot()
}
