% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bb_ea}
\alias{bb_ea}
\title{Euro Area budget balance as a percentage of GDP}
\format{
\subsection{\code{bb_ea}}{

A tibble object with 30 observations of 2 variables:
\describe{
\item{year}{a numeric indicating the year of the observation}
\item{value}{a numeric indicating the value of the budget balance for the Euro Area for the respective year (as a percentage of GDP)}
}
}
}
\source{
\url{https://fred.stlouisfed.org/series/GGNLBAEZA188N}

\url{https://ec.europa.eu/eurostat/databrowser/bookmark/72101904-17ca-46c3-accb-1d847055ca01?lang=en}
}
\usage{
bb_ea
}
\description{
Euro Area budget balance annual data from the International Monetary Fund and Eurostat measured as a percentage of GDP of the Euro Area for the respective year.
}
\details{
Net lending (+)/ borrowing (-) is calculated as revenue minus total expenditure. This is a core Government Finance Statistics (GFS) balance that measures the extent to which general government is either putting financial resources at the disposal of other sectors in the economy and nonresidents (net lending), or utilizing the financial resources generated by other sectors and nonresidents (net borrowing). This balance may be viewed as an indicator of the financial impact of general government activity on the rest of the economy and nonresidents (Government Finance Statistics Manual 2001 (GFSM 2001), paragraph 4.17).
}
\references{
International Monetary Fund, General government net lending/borrowing for Euro Area [GGNLBAEZA188N], retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/GGNLBAEZA188N, November 17, 2023.

Government deficit/surplus, debt and associated data [gov_10dd_edpt1] (2021), Eurostat, https://ec.europa.eu/eurostat/databrowser/bookmark/72101904-17ca-46c3-accb-1d847055ca01?lang=en, November 17, 2023.
}
\keyword{datasets}
