% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_gr_plot.R
\name{bb_gr_plot}
\alias{bb_gr_plot}
\title{Plot government budget balance during the Great Recession}
\usage{
bb_gr_plot()
}
\value{
plot object
}
\description{
This function plots the government budget balance (as a percentage of the respective GDP) for the USA and the Euro Area during the years of the Great Recession
and compares it to the baseline scenario of the Great Depression.
Since the Great Recession started at the end of 2007/beginning of 2008, the budget balance for the years 2008, 2009, and 2010 is plotted.
}
\examples{
bb_gr_plot()

}
