% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bb_us}
\alias{bb_us}
\title{US budget balance as a percentage of GDP}
\format{
\subsection{\code{bb_us}}{

A tibble object with 94 observations of 2 variables:
\describe{
\item{date}{a date indicating the year of the observation}
\item{value}{a numeric indicating the value of federal budget for the US for the respective year}
}
}
}
\source{
\url{https://fred.stlouisfed.org/series/FYFSGDA188S}
}
\usage{
bb_us
}
\description{
US budget balance annual data from the Federal Reserve Bank of St. Louis measured as a percentage of GDP.
The data covers the period from 1929 to 2022. Federal Surplus or Deficit [-] as Percent of Gross Domestic Product (FYFSGDA188S) was first constructed by the Federal Reserve Bank of St. Louis in October 2012. It is calculated using Federal Surplus or Deficit [-] (FYFSD) and Gross Domestic Product (GDPA):
}
\details{
FYFSGDA188S = ((FYFSD/1000)/GDPA)*100

FYFSD/1000 transforms FYFSD from millions of dollars to billions of dollars.
}
\references{
Federal Reserve Bank of St. Louis, Federal Surplus or Deficit \link{-} as Percent of Gross Domestic Product [FYFSGDA188S], retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/FYFSGDA188S, November 18, 2023.
}
\keyword{datasets}
