% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bitcoin}
\alias{bitcoin}
\title{Bitcoin price data}
\format{
\subsection{\code{bitcoin}}{

A tibble object with 3368 observations of 2 variables:
\describe{
\item{date}{a date indicating the date of the observation}
\item{value}{a numeric indicating the price of Bitcoin for the respective date (in US dollars)}
}
@source \url{https://finance.yahoo.com/quote/BTC-USD/history?p=BTC-USD}
}
}
\usage{
bitcoin
}
\description{
Bitcoin daily price data from Yahoo Finance! measured in US dollars. The data covers the period from 2014-09-17 to 2023-12-06.
}
\keyword{datasets}
