% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitcoin_price_plot.R
\name{bitcoin_price_plot}
\alias{bitcoin_price_plot}
\title{Plot bitcoin price}
\usage{
bitcoin_price_plot()
}
\value{
plot object
}
\description{
This function plots the bitcoin price from Sep 17, 2014 to Dec 06, 2023.
}
\examples{
bitcoin_price_plot()
}
