% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.R
\name{bp}
\alias{bp}
\title{Create barplot with three bars}
\usage{
bp(
  bar_1_data = NULL,
  bar_2_data = NULL,
  bar_3_data = NULL,
  y_ticks = 1,
  y_axis = TRUE,
  y_lim = c(-100, 100),
  mar_values = c(3, 4, 2, 1),
  names.arg = c("t+1", "t+2", "t+3")
)
}
\arguments{
\item{bar_1_data}{time series data for bar 1 (a named vector)}

\item{bar_2_data}{time series data for bar 2 (a named vector)}

\item{bar_3_data}{time series data for bar 3 (a named vector)}

\item{y_ticks}{tick distance for y-axis}

\item{y_axis}{logical, should y-axis be added?}

\item{y_lim}{numerical vector denoting y-axis limits}

\item{mar_values}{numerical vector denoting margins as required by \code{par()}}
}
\value{
a plot object
}
\description{
Create barplot with three bars
}
\examples{
x <- seq(-100, 50, by = 75)
y <- seq(-50, 50, by = 50)
z <- seq(-20, 60, by = 40)
bp(x, y, z, y_ticks = 20)

}
