% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_2.R
\name{bp_2}
\alias{bp_2}
\title{Create barplot with two adjacent bars}
\usage{
bp_2(
  bar_1_data = NULL,
  bar_2_data = NULL,
  y_ticks = 1,
  y_axis = TRUE,
  y_lim = c(-100, 100),
  mar_values = c(3, 4, 2, 1),
  names.arg = NULL
)
}
\arguments{
\item{bar_1_data}{time series data for bar 1 (a named vector)}

\item{bar_2_data}{time series data for bar 2 (a named vector)}

\item{y_ticks}{tick distance for y-axis}

\item{y_axis}{logical, should y-axis be added?}

\item{y_lim}{numerical vector denoting y-axis limits}

\item{mar_values}{numerical vector denoting margins as required by \code{par()}}
}
\value{
a plot object
}
\description{
Create barplot with two adjacent bars
}
\examples{
x <- seq(-100, 50, by = 75)
y <- seq(-50, 50, by = 50)
bp(x, y, y_ticks = 20)

}
