% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_covid_plot.R
\name{emp_covid_plot}
\alias{emp_covid_plot}
\title{Plot employment change during the COVID-19 pandemic}
\usage{
emp_covid_plot()
}
\value{
plot object
}
\description{
This function plots the change in employment compared to the base year (in percents) during the years of the COVID-19 pandemic for USA and the Euro Area
and compares it to the baseline scenario of the Great Depression.
Since the pandemic started at the beginning of 2022, the change in employment for the years 2020, 2021, and 2022 is plotted.
}
\examples{
emp_covid_plot()
}
