% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emp_ea}
\alias{emp_ea}
\title{Euro Area employment}
\format{
\subsection{\code{emp_ea}}{

A tibble object with 25 observations of 2 variables:
\describe{
\item{date}{a numeric indicating the year of the observation}
\item{value}{a numeric indicating the employed population in Euro Area for the respective year (in thousands of persons)}
}
@source \url{https://fred.stlouisfed.org/series/LFEMTTTTEZA647S}
@references Organization for Economic Co-operation and Development, Employed Population: Aged 15 and over: All Persons for Euro Area (19 Countries) [LFEMTTTTEZA647S], retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/LFEMTTTTEZA647S, November 23, 2023.
}
}
\usage{
emp_ea
}
\description{
Euro Area employment annual data from OECD, measured in thousands of persons. The data covers the period from 1998 to 2022.
}
\keyword{datasets}
