% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_gr_plot.R
\name{emp_gr_plot}
\alias{emp_gr_plot}
\title{Plot employment change during the Great Recession}
\usage{
emp_gr_plot()
}
\value{
plot object
}
\description{
This function plots the change in employment compared to the base year (in percents) during the years of the Great Recession for USA and the Euro Area
and compares it to the baseline scenario of the Great Depression.
Since the Great Recession started at the end of 2007/beginning of 2008, the change in employment for the years 2008, 2009, and 2010 is plotted.
}
\examples{
emp_gr_plot()

}
