% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_ea}
\alias{gdp_ea}
\title{Euro Area nominal GDP}
\format{
\subsection{\code{gdp_ea}}{

A tibble object with 28 observations of 2 variables:
\describe{
\item{year}{a numeric indicating the year of the observation}
\item{value}{a numeric indicating the gross domestic product for the Euro Area for the respective year (in millions of Euro/ECU)}
}
}
}
\source{
\url{https://fred.stlouisfed.org/series/CPMEURNSAB1GQEA19}
}
\usage{
gdp_ea
}
\description{
Euro Area nominal GDP annual data from Eurostat measured in millions of Euros/ECU. The data covers the period from 1995 to 2022.
}
\details{
Euro/ECU series is expressed in euro from January 1, 1999 till present. Prior to December 31, 1998, synthetic exchange rate of the national currency to European Community Unit (ECU) is used to adjust for market exchange rate movements.
}
\references{
UEurostat, Gross Domestic Product (Euro/ECU Series) for Euro Area (19 Countries) [CPMEURNSAB1GQEA19], retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/CPMEURNSAB1GQEA19, November 18, 2023.
}
\keyword{datasets}
