% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_us}
\alias{gdp_us}
\title{US nominal GDP}
\format{
\subsection{\code{gdp_us}}{

A tibble object with 94 observations of 2 variables:
\describe{
\item{year}{a numeric indicating the year of the observation}
\item{value}{a numeric indicating the gross domestic product for the US for the respective year (in billions of US dollars)}
}
}
}
\source{
\url{https://fred.stlouisfed.org/series/GDPA}
}
\usage{
gdp_us
}
\description{
US nominal GDP annual data from the U.S. Bureau of Economic Analysis, measured in billions of US dollars.
The data covers the period from 1929 to 2022.
}
\references{
U.S. Bureau of Economic Analysis, Gross Domestic Product \link{GDPA}, retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/GDPA, November 17, 2023.
}
\keyword{datasets}
