% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indpro_baseline.R
\name{indpro_baseline}
\alias{indpro_baseline}
\title{Baseline industrial production}
\usage{
indpro_baseline()
}
\value{
plot object
}
\description{
This function is plotting the dynamics of the US industrial production index for
the period Jul 1929 - Jul 1932 (peak to minimum). The resulting plot can be used
as a plot of baseline scenario and other time series can be added to it.
}
\examples{
indpro_baseline()
}
