% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indpro_covid_plot.R
\name{indpro_covid_plot}
\alias{indpro_covid_plot}
\title{Plot industrial production during the COVID-19 pandemic}
\usage{
indpro_covid_plot()
}
\value{
plot object
}
\description{
This function is plotting the dynamics of the US industrial production index for
the period Feb 2020 - Feb 2023 (peak to minimum).
}
\examples{
indpro_covid_plot()
}
