% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indpro_ea}
\alias{indpro_ea}
\title{FRED EA19 industrial production index}
\format{
\subsection{\code{indpro_ea}}{

A tibble object with 577 rows and 2 columns:
\describe{
\item{date}{a date indicating the year and the month in the time series}
\item{value}{a numeric indicating the value of the industrial production index for the respective month}
}
}
}
\source{
\url{https://fred.stlouisfed.org/series/EA19PRINTO01IXOBSAM}
}
\usage{
indpro_ea
}
\description{
Industrial production monthly data for the Euro Area (19 countries) from the OECD measured
as a seasonally-adjusted index with base year 2015. The original source of the data
is the release \href{http://www.oecd-ilibrary.org/economics/data/main-economic-indicators/main-economic-indicators-complete-database_data-00052-en}{Main Economic Indicators}.
}
\references{
Organization for Economic Co-operation and Development, Production: Industry: Total Industry Excluding Construction for Euro Area (19 Countries) [EA19PRINTO01IXOBSAM], retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/EA19PRINTO01IXOBSAM, November 13, 2023.
}
\keyword{datasets}
