% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indpro_gr.R
\name{indpro_gr}
\alias{indpro_gr}
\title{Industrial production during the Great Recession}
\usage{
indpro_gr()
}
\value{
plot object
}
\description{
Industrial production during the Great Recession
}
\examples{
indpro_gr()

}
