% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indpro_gr_plot.R
\name{indpro_gr_plot}
\alias{indpro_gr_plot}
\title{Industrial production during the Great Recession plot}
\usage{
indpro_gr_plot()
}
\value{
plot object
}
\description{
A function that plots industrial production in USA and the Euro Area during the Great Recession of 2008-2009
and compares it to the baseline scenario of the Great Depression of 1929-1933.
}
\examples{
indpro_gr_plot()

}
