% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_covid_plot_monthly.R
\name{mb_covid_plot_monthly}
\alias{mb_covid_plot_monthly}
\title{Plot monthly monetary base change during the COVID-19 pandemic}
\usage{
mb_covid_plot_monthly()
}
\value{
plot object
}
\description{
This function plots the monetary base change (as a percentage of GDP) during the years of the COVID-19 pandemic for USA and the Euro Area
and compares it to the baseline scenario of the Great Depression.
Since the pandemic started at the beginning of 2022, the change in monetary base for the years 2020, 2021, and 2022 is plotted.
}
\examples{
mb_covid_plot_monthly()
}
