% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_gr_plot.R
\name{mb_gr_plot}
\alias{mb_gr_plot}
\title{Plot monetary base change during the Great Recession}
\usage{
mb_gr_plot()
}
\value{
plot object
}
\description{
This function plots the change in the monetary base (as a percentage of the respective GDP) during the years of the Great Recession for USA and the Euro Area
and compares it to the baseline scenario of the Great Depression.
Since the Great Recession started at the end of 2007/beginning of 2008, the change in the monetary base for the years 2008, 2009, and 2010 is plotted.
}
\examples{
mb_gr_plot()

}
