% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mb_us}
\alias{mb_us}
\title{US monetary base}
\format{
\subsection{\code{mb_us}}{

A tibble object with 1269 observations of 2 variables:
\describe{
\item{date}{a date indicating the year and the month of the observation}
\item{value}{a numeric indicating the value of the monetary base for the respective month (in billions of US dollars)}
}
}
}
\source{
\url{https://fred.stlouisfed.org/series/BOGMBASE}

\url{https://fred.stlouisfed.org/series/AMBNS}
}
\usage{
mb_us
}
\description{
Monetary base monthly data from the Federal Reserve Bank of St. Louis measured
in billions of US dollars. The data cover the period from Jan 1918 to September 2023 and is not seasonally adjusted. The monetary base equals total reserve balances maintained by US banks plus currency in circulation.
It is created by merging two datasets from the Federal Reserve Bank of St. Louis:
\href{https://fred.stlouisfed.org/series/AMBNS}{AMBNS} and \href{https://fred.stlouisfed.org/series/BOGMBASE}{BOGMBASE}.
The first dataset ends on Nov 2019 and the second one is added to the time series from Dec 2019 to create a complete data set.
}
\references{
Board of Governors of the Federal Reserve System (US), Monetary Base; Total [BOGMBASE], retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/BOGMBASE, November 18, 2023.

Federal Reserve Bank of St. Louis, St. Louis Adjusted Monetary Base (DISCONTINUED) [AMBNS], retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/AMBNS, November 18, 2023.
}
\keyword{datasets}
