% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_figures.R
\name{print_presentation_figures}
\alias{print_presentation_figures}
\title{Print presentation figures}
\usage{
print_presentation_figures()
}
\value{
.tex files
}
\description{
Print presentation figures
}
\examples{
print_presentation_figures()

}
