% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_baseline.R
\name{stocks_baseline}
\alias{stocks_baseline}
\title{Stocks dynamics during the Great Depression plot}
\usage{
stocks_baseline()
}
\value{
plot object
}
\description{
This function is plotting the dynamics of the S&P 500 index for
the period Sep 1929 - Jun 1932 (peak to minimum). The resulting plot can be used
as a plot of baseline scenario and other time series can be added to it.
}
\examples{
stocks_baseline()
}
