% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_gr_plot.R
\name{stocks_gr_plot}
\alias{stocks_gr_plot}
\title{Stocks during the Great Recession plot}
\usage{
stocks_gr_plot(max_value = 50)
}
\value{
plot object
}
\description{
A function that plots stock indices in USA and the Euro Area during the Great Recession of 2008-2009
and compares it to the baseline scenario of the Great Depression of 1929-1933.
}
\examples{

stocks_gr_plot()
}
