% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mb_ea}
\alias{mb_ea}
\title{Euro Area monetary base (base money)}
\format{
\subsection{\code{mb_ea}}{

A tibble object with 261 observations of 2 variables:
\describe{
\item{date}{a date indicating the year and the month of the observation}
\item{value}{a numeric indicating the value of the monetary base (base money) at the end of the respective month (in millions of Euros)}
}
}
}
\source{
\url{https://data.ecb.europa.eu/data/datasets/ILM/ILM.M.U2.C.LT00001.Z5.EUR}
}
\usage{
mb_ea
}
\description{
Euro Area monetary base monthly data from the European Central Bank measured in millions of Euros. The data covers the period from Feb 1999 to Oct 2023.
}
\references{
European Central Bank, Frankfurt am Main, Germany. Base money, Euro area (changing composition), Monthly. https://data.ecb.europa.eu/data/datasets/ILM/ILM.M.U2.C.LT00001.Z5.EUR, November 2, 2023.
}
\keyword{datasets}
