% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_baseline_plot.R
\name{stocks_baseline_plot}
\alias{stocks_baseline_plot}
\title{Stocks dynamics during the Great Depression plot}
\usage{
stocks_baseline_plot(max_value = 0, y_axis = TRUE)
}
\arguments{
\item{max_value}{numeric providing the maximum value on the y-axis}
}
\value{
plot object
}
\description{
This function is plotting the dynamics of the S&P 500 index for
the period Sep 1929 - Jun 1932 (monthly peak to monthly minimum). The resulting plot can be used
as a plot of baseline scenario and other time series can be added to it.
}
\examples{
stocks_baseline_plot()
}
